refactor: update main handler with run.py
